DataBase/requirements.txt
SQLAlchemy==1.4.23
alembic==1.7.1
psycopg2-binary==2.9.1 